use "unsigned long" for max_long_transfer

+++ src/libopenusb1-glue.c
@@ -1256,7 +1256,7 @@ ptp_usb_senddata(PTPParams* params, PTPContainer* ptp,
     bytes_left_to_transfer = size - datawlen;
     ret = PTP_RC_OK;
     while (bytes_left_to_transfer > 0) {
-	int max_long_transfer = ULONG_MAX + 1 - packet_size;
+	unsigned long max_long_transfer = ULONG_MAX + 1 - packet_size;
 	ret = ptp_write_func (bytes_left_to_transfer > max_long_transfer ? max_long_transfer : bytes_left_to_transfer,
 		handler, params->data, &written);
         if (ret != PTP_RC_OK){
